requirements.txt
certifi==2017.11.5
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.10
psycopg2==2.7.3.2
pytz==2017.3
wincertstore==0.2
